displays score and time in the top left

+++ Game2Boat/src/test/GameTest.java
@@ -11,23 +11,23 @@ public class GameTest {
 	@Test
 	public void scoreTest() {
 		Game g1 = new Game();
-		assertEquals(g1.getScore(), 0);
+		assertEquals((int)g1.getScore(), 0);
 		g1.decreaseScore(1);
-		assertEquals(g1.getScore(), 0); // check underflow
+		assertEquals((int)g1.getScore(), 0); // check underflow
 		g1.increaseScore(1);
-		assertEquals(g1.getScore(), 1);
+		assertEquals((int)g1.getScore(), 1);
 		g1.decreaseScore(1);
-		assertEquals(g1.getScore(), 0);
+		assertEquals((int)g1.getScore(), 0);
 	}
 	@Test
 	public void timeTest(){
 		Game g1 = new Game();
 		for (int i = 120; i >= 0; i--){
-			assertEquals(g1.getTime(), i);
+			assertEquals((int)g1.getTime(), i);
 			g1.decreaseTime();
 		}
 		g1.decreaseTime();
-		assertEquals(g1.getTime(), 0);  // check underflow
+		assertEquals((int)g1.getTime(), 0);  // check underflow
 	}
 	@Test
 	public void lapTest(){
